organize lidars into their own package
