+++ unified_frameworks/sensor_array/lidar/lidar.py
@@ -2,11 +2,11 @@ import sys
 import re
 root = (next(re.finditer(".*unified_frameworks", __file__)).group())
 sys.path.append(root) if root not in sys.path else None
-from sensor_array.actual_lidar import ActualLidar
-from sensor_array.bridge_lidar import BridgeLidar
+from unified_frameworks.sensor_array.lidar.actual_lidar import ActualLidar
+from unified_frameworks.sensor_array.lidar.bridge_lidar import BridgeLidar
 import importlib
-from sensor_array.fake_lidar import FakeLidar
-from sensor_array.LidarClass import _Lidar
+from unified_frameworks.sensor_array.lidar.fake_lidar import FakeLidar
+from unified_frameworks.sensor_array.lidar.LidarClass import _Lidar
 import traceback
 from threading import Thread
 from math import pi, cos, sin, sqrt, atan2, radians
