+++ unified_frameworks/worldview.py
@@ -7,12 +7,12 @@ import time
 from threading import Thread
 from math import pi
 from unified_utils import polar_sum, _Abstract_Service
-import sensor_array.lidar
+import unified_frameworks.sensor_array.lidar.lidar
 import bridge.client_side
 import importlib
 importlib.reload(bridge.client_side)
-importlib.reload(sensor_array.lidar)
-lidar = sensor_array.lidar
+importlib.reload(unified_frameworks.sensor_array.lidar.lidar)
+lidar = unified_frameworks.sensor_array.lidar.lidar
 
 
 config = {
